+++ projects/com.oracle.truffle.llvm.nodes/src/com/oracle/truffle/llvm/nodes/intrinsics/interop/LLVMTruffleAddressToFunction.java
@@ -55,7 +55,7 @@ public abstract class LLVMTruffleAddressToFunction extends LLVMIntrinsic {
     }
 
     protected static LLVMFunctionDescriptor getDescriptor(LLVMAddress value, LLVMContext cachedContext) {
-        return cachedContext.lookup(new LLVMFunctionHandle((int) value.getVal()));
+        return cachedContext.lookup(new LLVMFunctionHandle(value.getVal()));
     }
 
 }
