+++ projects/com.oracle.truffle.llvm.runtime/src/com/oracle/truffle/llvm/runtime/memory/LLVMHeap.java
@@ -48,13 +48,13 @@ public final class LLVMHeap extends LLVMMemory {
     // current hack: we cannot directly store the LLVMFunction in the native memory due to GC
     public static final int FUNCTION_PTR_SIZE_BYTE = 8;
 
-    public static void putFunctionIndex(LLVMAddress address, int functionIndex) {
-        UNSAFE.putInt(LLVMMemory.extractAddr(address), functionIndex);
+    public static void putFunctionIndex(LLVMAddress address, long functionIndex) {
+        UNSAFE.putLong(LLVMMemory.extractAddr(address), functionIndex);
     }
 
-    public static void putFunctionIndex(long ptr, int functionIndex) {
+    public static void putFunctionIndex(long ptr, long functionIndex) {
         assert ptr != 0;
-        UNSAFE.putInt(ptr, functionIndex);
+        UNSAFE.putLong(ptr, functionIndex);
     }
 
     public static int getFunctionIndex(LLVMAddress addr) {
